refactor: float -> double 변경
- 주류 도수 타입 double로 변경
- 리뷰 총합 타입 double로 변경

+++ src/main/java/org/complete/challang/drink/domain/entity/Drink.java
@@ -25,13 +25,13 @@ public class Drink extends BaseEntity {
 
     private String description;
 
-    private float abv;
+    private double abv;
 
     private String imageUrl;
 
     private long reviewCount;
 
-    private float reviewSumRating;
+    private double reviewSumRating;
 
     @Embedded
     private FoodStatistic foodStatistic;
